requirements.txt
Arpeggio==1.9.0
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
decorator==4.3.0
Faker==1.0.2
graphviz==0.10.1
idna==2.8
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
Jinja2==2.10
kafka-python==1.4.4
MarkupSafe==1.1.0
numpy==1.16.2
parglare==0.8.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pudb==2018.1
py4j==0.10.7
Pygments==2.2.0
pyspark==2.4.0
pyspark-stubs==2.4.0.post1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
rfc3339==6.0
six==1.11.0
text-unidecode==1.2
textX==1.8.0
traitlets==4.3.2
tzlocal==1.5.1
urllib3==1.24.1
urwid==2.0.1
wcwidth==0.1.7
